drivers.json: add aliases node for serial UART
The aliases node is needed for spinal_lib_uart driver to get the ID
for the UART devices. This ID will be used during UART driver registration.
If the aliases node not exists, UART driver will not able to register.
As a result, user not able to login or use the serial UART.

Signed-off-by: Mohamad Noor Alim Hussin <mnalim@efinixinc.com>