DT: refactor code for creating memory node
DT for Zephyr will create 2 memory nodes when on chip ram
and external memory are enable. As a result, the last entry
of the Python dictionary will overwrite when the dictionary key
of "memory" match with the existing dictionary. To resolve
this issue, the memory node for Zephyr has been renamed to memory0
and memory1 for on chip ram and external memory respectively.

Signed-off-by: Mohamad Noor Alim Hussin <mnalim@efinixinc.com>