core.py: add a function to get bus address
The keyword for bus do not follow the peripheral keyword pattern.
Add a few helper function to filter the bus keyword and grep the
address of the bus. These helper function is require to get the
correct address of the bus else it may accidentaly get the wrong
value such as size of the bus.

Signed-off-by: Mohamad Noor Alim Hussin <mnalim@efinixinc.com>