DT: user_config argument support adding child node
Add support for adding child node using user_config argument
in a single file. The original user_config is only to override
the peripheral properties. However, if user want to override
and add child node for that peripheral, it is better to make
it in a single file.

For example, user want to override the compatible property
for i2c0 master and add a slave device called bmp280, they
could do as follows.

{
	"override": {
		"I2C_0": {
			"compatible": "i2c-gpio"
		}
	},
	"child": {
		"child1": {
			"parent_label": "i2c0",
			"name": "bmp280",
			"label": "bmp280",
			"address": 0,
			"compatible": "bosch,bmp280",
			"status": "okay"
		}
	}
}

Signed-off-by: Mohamad Noor Alim Hussin <mnalim@efinixinc.com>