core.py: fix calculate offset address of peripheral with dual buses system
With dual buses system, the peripherals are connected to different
bus address. The issue arise due to peripheral that connected to
second bus use the first bus address as a base address.

For example, spi1 peripheral is connected to bus1 and spi2 is connected
to bus2. The address offset of spi2 is calculated based on bus1 instead
of bus2.

Signed-off-by: Mohamad Noor Alim Hussin <mnalim@efinixinc.com>