bus_node.py: add method to update the header and reg property
The Linux device tree is using address offset in reg property.
When a device node created using create_node(), the reg and
header is using the actual address value. The address offset
only available if the device node is connect to the bus node.
Once we know the bus node starting address, we could calculate
the device address offset and update the header and reg property.

Signed-off-by: Mohamad Noor Alim Hussin <mnalim@efinixinc.com>