bus_node.py: handle generating reg and header if address offset is false
The size-cells is hardcoded to 0 when populating device nodes.
This causing the reg produce the wrong output. However, if the size-cells
is set to 1, it would produce the correct output but with wrong format.

We could override the size-cells for each device in user-defined configuration
but it would tedious. The simpler way is to regenerate the reg.

Signed-off-by: Mohamad Noor Alim Hussin <mnalim@efinixinc.com>