bus_node.py: allow user to override reg property in json file
Current implementation automatically generate reg property based on
address mapping and size. It also does not support reg property with
multiple values. For example, user might have a reg property in device
node as follows,

	reg = <0x1000 0x100, 0x1100 0x100>;

For this case, we allow user to override the reg property in the
user configuration json file as follows.

{
    "overrides": {
        "dev1": {
            "reg": "0x1000 0x100, 0x1100 0x100",
        }
    }
}

Signed-off-by: Mohamad Noor Alim Hussin <mnalim@efinixinc.com>